# Improve performance of statistics