- Close #3750

Running with `--statistics` is now faster than without it.

```console
❯ hyperfine 'cargo run -p ruff_cli -- check --no-cache --statistics --select=ALL . || true'
Benchmark 1: cargo run -p ruff_cli -- check --no-cache --statistics --select=ALL . || true
  Time (mean ± σ):     749.5 ms ±  17.1 ms    [User: 4181.3 ms, System: 100.7 ms]
  Range (min … max):   721.1 ms … 774.2 ms    10 runs
 
❯ hyperfine 'cargo run -p ruff_cli -- check --no-cache --select=ALL . || true'
Benchmark 1: cargo run -p ruff_cli -- check --no-cache --select=ALL . || true
  Time (mean ± σ):     792.6 ms ±  16.4 ms    [User: 4219.7 ms, System: 94.8 ms]
  Range (min … max):   772.9 ms … 829.8 ms    10 runs
```