@JonathanPlasse 2023-03-26
```console
# Before
❯ hyperfine 'cargo run -p ruff_cli -- check --no-cache --statistics --select=ALL . || true'
Benchmark 1: cargo run -p ruff_cli -- check --no-cache --statistics --select=ALL . || true
  Time (mean ± σ):     752.7 ms ±  10.7 ms    [User: 4186.4 ms, System: 100.1 ms]
  Range (min … max):   736.7 ms … 769.4 ms    10 runs

# After
❯ hyperfine 'cargo run -p ruff_cli -- check --no-cache --statistics --select=ALL . || true'
Benchmark 1: cargo run -p ruff_cli -- check --no-cache --statistics --select=ALL . || true
  Time (mean ± σ):     739.4 ms ±  14.6 ms    [User: 4179.7 ms, System: 92.5 ms]
  Range (min … max):   723.9 ms … 763.4 ms    10 runs
```